code/Stack/ViewModel/src/test/java/ru/unn/agile/Stack/viewmodel/ViewModelTests.java
package ru.unn.agile.Stack.viewmodel;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class ViewModelTests {
    private ViewModel viewModel;

    @Before
    public void setUp() {
        viewModel = new ViewModel();
    }

    @After
    public void tearDown() {
        viewModel = null;
    }

    @Test
    public void canSetDefaultValues() {
        assertEquals("", viewModel.txtinputProperty().get());
        assertEquals("", viewModel.txtprintProperty().get());
        assertEquals("", viewModel.txtlogProperty().get());
        assertEquals("", viewModel.txtmsgProperty().get());
    }

    @Test
    public void badTyping() {
        viewModel.txtinputProperty().set("a");
        viewModel.push();
        assertEquals("Bad input", viewModel.txtlogProperty().get());
    }
}
